Documented and tested DataPoint class
