+++ test/utility_test.cpp
@@ -2,7 +2,7 @@
  * Tests for the utility.h header. Includes helper methods for dealing with
  * sparse matrices.
  *
- * This file is provided as part of the STAG library and released under the MIT
+ * This file is provided as part of the STAG library and released under the GPL
  * license.
  */
 #include <gtest/gtest.h>
